refactor jsx functions location
